+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.43.0...HEAD)
 
+Updated environments:
+
+- **devon_rex** 2.40.4 -> 2.40.5 [#2102](https://github.com/sider/runners/pull/2102)
+
 Updated tools:
 
 - **GolangCI-Lint** 1.36.0 -> 1.37.1 [#2086](https://github.com/sider/runners/pull/2086)
@@ -15,6 +19,8 @@ Misc:
 
 - **Flake8** Provide Sider's recommended ruleset [#1980](https://github.com/sider/runners/pull/1980)
 - **PHP_CodeSniffer** Provide Sider's recommended ruleset [#2084](https://github.com/sider/runners/pull/2084)
+- Remove *removed* Go tools from `sider.yml` [#2099](https://github.com/sider/runners/pull/2099)
+- Add `Runners::ConfigGenerator` class [#2029](https://github.com/sider/runners/pull/2029)
 - **Metrics FileInfo** Modify target file generation [#2100](https://github.com/sider/runners/pull/2100)
 
 ## 0.43.0